added a layer for working with the db and processing all incoming messages.
