+++ src/main/java/com/dev/vlpr/service/MainService.java
@@ -1,0 +1,7 @@
+package com.dev.vlpr.service;
+
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+public interface MainService {
+    void processTextMessage(Update update);
+}
