+++ src/main/java/com/dev/vlpr/service/impl/ConsumerServiceImpl.java
@@ -1,11 +1,10 @@
 package com.dev.vlpr.service.impl;
 
 import com.dev.vlpr.service.ConsumerService;
-import com.dev.vlpr.service.ProducerService;
+import com.dev.vlpr.service.MainService;
 import lombok.extern.log4j.Log4j;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Service;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
 import static com.dev.vlpr.common.RabbitQueue.*;
@@ -14,22 +13,17 @@ import static com.dev.vlpr.common.RabbitQueue.*;
 @Log4j
 public class ConsumerServiceImpl implements ConsumerService {
 
-    private final ProducerService producerService;
+    private final MainService mainService;
 
-    public ConsumerServiceImpl(ProducerService producerService) {
-        this.producerService = producerService;
+    public ConsumerServiceImpl(MainService mainService) {
+        this.mainService = mainService;
     }
 
     @Override
     @RabbitListener(queues = TEXT_MESSAGE_UPDATE)
     public void consumerTextMessageUpdate(Update update) {
         log.debug("node: txt message is received");
-
-        var message = update.getMessage();
-        var sendMessage = new SendMessage();
-        sendMessage.setChatId(message.getChatId().toString());
-        sendMessage.setText("hi from node");
-        producerService.producerAnswer(sendMessage);
+        mainService.processTextMessage(update);
     }
 
     @Override
